requirements.txt
Flask>=2.0
google-cloud-storage>=2.0
Werkzeug>=2.0
python-dotenv>=0.19
protobuf~=6.30.2
MarkupSafe~=3.0.2
